[이슈 #7] 디테일 액티비티 레이아웃
