+++ Frontend/HackPretty/app/src/main/AndroidManifest.xml
@@ -25,10 +25,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-<<<<<<< HEAD
-        <activity android:name=".ListActivity" />
-=======
->>>>>>> fb4dcd36f421c9884a7f3f8bfe46a31fd0a23c6b
+
         <activity android:name=".DetailActivity"></activity>
     </application>
