+++ Frontend/HackPretty/app/src/main/java/com/example/haeun_kim/hackpretty/SearchActivity.java
@@ -8,13 +8,15 @@ import android.provider.MediaStore;
 import android.support.v7.app.AlertDialog;
 import android.util.Log;
 import android.widget.Button;
-
+import android.widget.RelativeLayout;
 
 
 public class SearchActivity extends BaseActivity {
 
     private static final int PICK_FROM_CAMERA = 1;
     private static final int PICK_FROM_GALLERY = 2;
+
+    RelativeLayout relativeLayout;
 
 
     @Override
@@ -24,6 +26,9 @@ public class SearchActivity extends BaseActivity {
 
         Button btnSearch = (Button) findViewById(R.id.btnSearch);
         Button btnCamera = (Button) findViewById(R.id.btnCamera);
+
+
+
 
 
         btnSearch.setOnClickListener((v) -> {
@@ -69,6 +74,8 @@ public class SearchActivity extends BaseActivity {
             }
         }
 
+
+        //DetailActivity로 Intent전달
         Intent intent = new Intent(SearchActivity.this, DetailActivity.class);
         startActivity(intent);
