+++ Frontend/HackPretty/app/src/main/res/layout/activity_search.xml
@@ -4,8 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/background"
+    android:id="@+id/searchLayout"
     tools:context="com.example.haeun_kim.hackpretty.SearchActivity">
+
 
     <LinearLayout
         android:layout_width="match_parent"
